# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KPI Command Center",
    "description": "Operaciones ATM, sucursales y planificación geoespacial. Importes en DOP.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/atms": {
      "get": {
        "tags": [
          "Cajeros automáticos"
        ],
        "summary": "List Atms",
        "operationId": "list_atms_api_v1_atms_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "estado",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ATMStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Estado"
            }
          },
          {
            "name": "provincia",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provincia"
            }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "name": "tipo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ATMType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tipo"
            }
          },
          {
            "name": "bajo_efectivo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bajo Efectivo"
            }
          },
          {
            "name": "sucursal_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sucursal Id"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ATMOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Cajeros automáticos"
        ],
        "summary": "Create Atm",
        "operationId": "create_atm_api_v1_atms_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ATMCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ATMOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/atms/{id}": {
      "get": {
        "tags": [
          "Cajeros automáticos"
        ],
        "summary": "Detail",
        "operationId": "detail_api_v1_atms__id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ATMOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/atms/{id}/status": {
      "patch": {
        "tags": [
          "Cajeros automáticos"
        ],
        "summary": "Patch Status",
        "operationId": "patch_status_api_v1_atms__id__status_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ATMStatusPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ATMOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/atms/{id}/heartbeat": {
      "post": {
        "tags": [
          "Cajeros automáticos"
        ],
        "summary": "Heartbeat",
        "operationId": "heartbeat_api_v1_atms__id__heartbeat_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Heartbeat"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ATMOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/atms/{id}/readings": {
      "post": {
        "tags": [
          "Cajeros automáticos"
        ],
        "summary": "Ingest Reading",
        "operationId": "ingest_reading_api_v1_atms__id__readings_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ATMReadingCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ATMReadingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Cajeros automáticos"
        ],
        "summary": "Readings",
        "operationId": "readings_api_v1_atms__id__readings_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ATMReadingOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/branches/{id}/atms": {
      "get": {
        "tags": [
          "Sucursales y colas"
        ],
        "summary": "Branch Atms",
        "operationId": "branch_atms_api_v1_branches__id__atms_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ATMOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ATMCreate": {
        "properties": {
          "codigo_unico": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Codigo Unico"
          },
          "sucursal_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sucursal Id"
          },
          "nombre": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Nombre"
          },
          "provincia": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Provincia"
          },
          "municipio": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Municipio"
          },
          "direccion": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Direccion"
          },
          "tipo": {
            "$ref": "#/components/schemas/ATMType"
          },
          "nivel_efectivo_pct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Nivel Efectivo Pct",
            "default": 100
          },
          "capacidad_efectivo": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,16}|(?=[\\d.]{1,19}0*$)\\d{0,16}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Capacidad Efectivo",
            "default": "1000000.00"
          },
          "estado": {
            "$ref": "#/components/schemas/ATMStatus",
            "default": "OPERATIVO"
          },
          "latitud": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitud"
          },
          "longitud": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitud"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "codigo_unico",
          "nombre",
          "provincia",
          "municipio",
          "direccion",
          "tipo",
          "latitud",
          "longitud"
        ],
        "title": "ATMCreate"
      },
      "ATMOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "codigo_unico": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Codigo Unico"
          },
          "sucursal_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sucursal Id"
          },
          "nombre": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Nombre"
          },
          "provincia": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Provincia"
          },
          "municipio": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Municipio"
          },
          "direccion": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Direccion"
          },
          "tipo": {
            "$ref": "#/components/schemas/ATMType"
          },
          "nivel_efectivo_pct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Nivel Efectivo Pct",
            "default": 100
          },
          "capacidad_efectivo": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,16}|(?=[\\d.]{1,19}0*$)\\d{0,16}\\.\\d{0,2}0*$)",
            "title": "Capacidad Efectivo",
            "default": "1000000.00"
          },
          "estado": {
            "$ref": "#/components/schemas/ATMStatus",
            "default": "OPERATIVO"
          },
          "latitud": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitud"
          },
          "longitud": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitud"
          },
          "ultima_comunicacion": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ultima Comunicacion"
          },
          "ultimo_mantenimiento": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ultimo Mantenimiento"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "codigo_unico",
          "nombre",
          "provincia",
          "municipio",
          "direccion",
          "tipo",
          "latitud",
          "longitud",
          "ultima_comunicacion",
          "ultimo_mantenimiento",
          "created_at",
          "updated_at"
        ],
        "title": "ATMOut"
      },
      "ATMReadingCreate": {
        "properties": {
          "hora": {
            "type": "string",
            "format": "date-time",
            "title": "Hora"
          },
          "transacciones": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Transacciones"
          },
          "transacciones_exitosas": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Transacciones Exitosas"
          },
          "monto_retirado": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,16}|(?=[\\d.]{1,19}0*$)\\d{0,16}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Monto Retirado"
          },
          "nivel_efectivo_pct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Nivel Efectivo Pct"
          },
          "segundos_observados": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 1.0,
            "title": "Segundos Observados"
          },
          "segundos_disponible": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 0.0,
            "title": "Segundos Disponible"
          },
          "segundos_utilizado": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 0.0,
            "title": "Segundos Utilizado"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "hora",
          "transacciones",
          "transacciones_exitosas",
          "monto_retirado",
          "nivel_efectivo_pct",
          "segundos_observados",
          "segundos_disponible",
          "segundos_utilizado"
        ],
        "title": "ATMReadingCreate"
      },
      "ATMReadingOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "hora": {
            "type": "string",
            "format": "date-time",
            "title": "Hora"
          },
          "transacciones": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Transacciones"
          },
          "transacciones_exitosas": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Transacciones Exitosas"
          },
          "monto_retirado": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,16}|(?=[\\d.]{1,19}0*$)\\d{0,16}\\.\\d{0,2}0*$)",
            "title": "Monto Retirado"
          },
          "nivel_efectivo_pct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Nivel Efectivo Pct"
          },
          "segundos_observados": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 1.0,
            "title": "Segundos Observados"
          },
          "segundos_disponible": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 0.0,
            "title": "Segundos Disponible"
          },
          "segundos_utilizado": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 0.0,
            "title": "Segundos Utilizado"
          },
          "atm_id": {
            "type": "string",
            "format": "uuid",
            "title": "Atm Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "hora",
          "transacciones",
          "transacciones_exitosas",
          "monto_retirado",
          "nivel_efectivo_pct",
          "segundos_observados",
          "segundos_disponible",
          "segundos_utilizado",
          "atm_id"
        ],
        "title": "ATMReadingOut"
      },
      "ATMStatus": {
        "type": "string",
        "enum": [
          "OPERATIVO",
          "BAJO_EFECTIVO",
          "FUERA_DE_SERVICIO",
          "MANTENIMIENTO"
        ],
        "title": "ATMStatus"
      },
      "ATMStatusPatch": {
        "properties": {
          "estado": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ATMStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "nivel_efectivo_pct": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Nivel Efectivo Pct"
          },
          "ultimo_mantenimiento": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ultimo Mantenimiento"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ATMStatusPatch"
      },
...
      "ATMType": {
        "type": "string",
        "enum": [
          "DISPENSADOR",
          "CDM_DEPOSITO"
        ],
        "title": "ATMType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Heartbeat": {
        "properties": {
          "estado": {
            "$ref": "#/components/schemas/ATMStatus"
          },
          "nivel_efectivo_pct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Nivel Efectivo Pct"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "estado",
          "nivel_efectivo_pct"
        ],
        "title": "Heartbeat"
      },
...
      "Page_ATMOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ATMOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ATMOut]"
      },
      "Page_ATMReadingOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ATMReadingOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ATMReadingOut]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}